more mess

+++ app/src/main/kotlin/com/stevenfrew/beatprompter/cache/AudioFile.kt
@@ -1,4 +1,4 @@
 package com.stevenfrew.beatprompter.cache
 
 @CacheXmlTag("audiofile")
-class AudioFile internal constructor(cachedCloudFileDescriptor: CachedCloudFileDescriptor, val mDurationMilliseconds:Int) : CachedCloudFile(cachedCloudFileDescriptor)
+class AudioFile internal constructor(cachedCloudFileDescriptor: CachedCloudFileDescriptor, val mDurationMilliseconds:Long) : CachedCloudFile(cachedCloudFileDescriptor)
